Split query stats into its own file
