+++ src/librustc/ty/query/mod.rs
@@ -52,7 +52,6 @@ use rustc_target::spec::PanicStrategy;
 use rustc_attr as attr;
 use rustc_span::symbol::Symbol;
 use rustc_span::{Span, DUMMY_SP};
-use std::any::type_name;
 use std::borrow::Cow;
 use std::convert::TryFrom;
 use std::ops::Deref;
@@ -63,6 +62,9 @@ use syntax::ast;
 mod plumbing;
 use self::plumbing::*;
 pub use self::plumbing::{force_from_dep_node, CycleError};
+
+mod stats;
+pub use self::stats::print_stats;
 
 mod job;
 #[cfg(parallel_compiler)]
